backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
django-redis>=5.4,<6.0
daphne>=4.0,<5.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
requests>=2.31,<3.0
